test: split connector ui coverage
